--- rosetta_eval/__init__.py
@@ -1,3 +1,0 @@
-from .tasks.medmcqa.medmcqa_task import *
-from .template_loader import *
-from .utils import *
