--- rosetta_eval/loader/__init__.py
@@ -1,1 +1,0 @@
-from rosetta_eval.loader.factory import load_dataset
