--- rosetta_eval/tasks/medmcqa/cot/__init__.py
@@ -1,1 +1,0 @@
-
