--- rosetta_eval/tasks/medmcqa/cot/prompt.jinja
@@ -1,18 +1,0 @@
-{% if description %}{{ description }}
-{% endif %}{% if instruction %}{{ instruction }}
-{% else %}The following are multiple choice questions about medical knowledge. Solve them in a step-by-step fashion, starting by summarizing the available information. Output a single option from the four options as the final answer.
-{% endif %}{% if examples %}{% for item in examples %}
-Question: {{ item.question }}
-Choices:{% for choice in item.choices %}
-{{ loop.index0 | to_letter }}. {{ choice }}{% endfor %}{% for key, value in item.items() %}{% if key not in ['question', 'choices', 'answer'] %}
-{{ key|capitalize }}: {{ value }}{% endif %}{% endfor %}
-Answer: {{ item.answer }}
-{% endfor %}{% endif %}
-Question: {{ question }}
-Choices:{% for choice in choices %}
-{{ loop.index }}. {{ choice }}{% endfor %}
-Explanation: 
-Answer: {{ answer }}
-
-
-
