--- rosetta_eval/tasks/medmcqa/ensemble_refinement/__init__.py
@@ -1,1 +1,0 @@
-
