--- rosetta_eval/tasks/medmcqa/few_shots/__init__.py
@@ -1,1 +1,0 @@
-
