--- rosetta_eval/tasks/medmcqa/few_shots/prompt.jinja
@@ -1,13 +1,0 @@
-{% if description %}{{ description }}
-{% endif %}{% if instruction %}{{ instruction }}
-{% else %}Please select the correct answer from the choices provided for each question.
-{% endif %}{% if examples %}{% for item in examples %}
-Question: {{ item.question }}
-Choices:{% for choice in item.choices %}
-{{ loop.index0 | to_letter }}. {{ choice }}{% endfor %}
-Answer: {{ item.answer }}
-{% endfor %}{% endif %}
-Question: {{ question }}
-Choices:{% for choice in choices %}
-{{ loop.index }}. {{ choice }}{% endfor %}
-Answer: {{ answer }}
