--- rosetta_eval/tasks/medmcqa/self_consistency/__init__.py
@@ -1,1 +1,0 @@
-
