+++ rosettaeval/loader/__init__.py
@@ -1,0 +1,1 @@
+from rosettaeval.loader.factory import load_dataset
