+++ rosettaeval/loader/tasks/medmcqa/__init__.py
@@ -1,0 +1,1 @@
+from rosettaeval.loader.tasks.medmcqa.base import MedMcQaQueryAdapter
