+++ rosettaeval/loader/tasks/medqa/__init__.py
@@ -1,0 +1,1 @@
+from rosettaeval.loader.tasks.medqa.base import MedQaQueryAdapter
